% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.module.representatives.R
\name{calculate.module.representatives}
\alias{calculate.module.representatives}
\title{Calculate Module Representative}
\usage{
calculate.module.representatives(data, representative.method = "mean")
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}} with three columns: name, sampleID and z-score.}

\item{representative.method}{the method, that is used for calculation of the representatives.
Currently implemented:"eigenM-PCA", "eigenM-SVD" and "mean"}
}
\value{
a \code{\link[data.table]{data.table}} containing the module representative for each sample
}
\description{
Calculate Module Representative
}
\references{
\insertRef{Langfelder2007}{MoDentify}

\insertRef{Chuang2007}{MoDentify}
}
