% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateModuleRepresentatives.R
\name{calculateModuleRepresentatives}
\alias{calculateModuleRepresentatives}
\title{Calculate Module Representative}
\usage{
calculateModuleRepresentatives(data, representative.method = "average")
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}} with three columns: name,
sampleID and z-score.}

\item{representative.method}{the method, that is used for calculation of the
representatives. Currently implemented:"eigenmetabolite" and "average"}
}
\value{
a \code{\link[data.table]{data.table}} containing the module
representative for each sample
}
\description{
Calculate Module Representative
}
\references{
\insertRef{Langfelder2007}{MoDentify}

\insertRef{Chuang2007}{MoDentify}
}
