% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateModuleScore.R
\name{calculateModuleScore}
\alias{calculateModuleScore}
\title{Calculate Module-score}
\usage{
calculateModuleScore(graph, nodes, data, phenotype, covars = NULL,
  representative.method = "average")
}
\arguments{
\item{graph}{an igraph object, which can be generated with \code{\link{generateNetwork}}.
The ID of the nodes must correspond to the name of the variables.}

\item{nodes}{a vector containing the ID of the nodes contained in the module}

\item{data}{a \code{\link[data.table]{data.table}} with three columns: name, sampleID and z-score.}

\item{phenotype}{a vector with the values for a phenotype of interest.
It must have the same number of observations as in data.}

\item{covars}{a \code{\link[data.table]{data.table}} containing the covariates.
The rows for the observations must be in the same order as in the phenotype vector.}

\item{representative.method}{the method, that is used for the calculation of the eigenmetabolites.
Currently implemented: "eigenmetabolite" and "average"}
}
\value{
a list containing the module score and the regression coefficient
}
\description{
Calculate Module-score
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)
data(qmdiab.phenos)

net.graph <- generateNetwork(data = qmdiab.data, annotations = qmdiab.annos)

data <- data.table(
  sampleID = paste0("sample", 1:dim(qmdiab.data)[1]),
  qmdiab.data
)
data <- melt(data = data, id.vars = "sampleID", variable.name = "name")
data[, z.score := scale(value), by = .(name)]

module.nodes <- c(3, neighbors(net.graph, 3))

module.score <- calculateModuleScore(
  graph = net.graph, nodes = module.nodes,
  data = data, phenotype = qmdiab.phenos$T2D
)
}
\references{
\insertRef{Do2017}{MoDentify}

\insertRef{Chuang2007}{MoDentify}
}
