% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateModuleScore.R
\name{calculateModuleScoreM}
\alias{calculateModuleScoreM}
\title{Calculate Module-score Memoized}
\usage{
calculateModuleScoreM(..., envir = parent.frame())
}
\description{
Calculate Module-score Memoized
}
\keyword{internal}
