% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.pathways.network.R
\name{create.pathways.network}
\alias{create.pathways.network}
\title{Subnetwork Creation function}
\usage{
create.pathways.network(data, covars = NULL, annotations,
  column.name = "Sub.pathway", alpha = 0.05,
  correction.method = "bonferroni", rm.unknown = TRUE,
  representative.method = "eigenM-PCA")
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}} or matrix containing data.
The columns correspond to the variables, the rows to the observations.}

\item{covars}{a \code{\link[data.table]{data.table}} containing covariates to correct for.
The columns correspond to the different covariates, the rows to the observations.}

\item{annotations}{a \code{\link[data.table]{data.table}} containing annotations for the variables.
The columns correspond to the different annotations, the rows to the variables.}

\item{column.name}{name of the column that should be used for grouping the variables. The default is "Sub.pathway".}

\item{alpha}{significance threshold (type 1 error) for multiple testing correction.}

\item{correction.method}{the method that should be used for multiple testing correction ("bonferroni", "BH", "BY", "fdr", "holm", "hochberg", "hommel", "none").
Default is bonferroni. See \code{\link[stats]{p.adjust}}.}

\item{rm.unknown}{remove variables, for which the group is unknown. The corresponding grouping variable name should be "Unknown".
Default is \code{TRUE}.}

\item{representative.method}{the method, that is used for calculation of the eigenmetabolites.
Currently implemented: "eigenM-PCA", "eigenM-SVD" and "mean"}
}
\value{
a list with a network containing the variables as nodes as
an \code{\link[igraph]{igraph}} object and a list with the module representatives
including their explained variances.
}
\description{
Subnetwork Creation function
}
\examples{
pathway.graph<-create.pathways.network(data = qmdiab.data,
 annotations = qmdiab.annotations, column.name = "Sub.pathway")
}
\references{
\insertRef{Krumsiek2011}{MoDentify}

\insertRef{Do2017}{MoDentify}
}
