% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.duplicates.R
\name{delete.duplicates}
\alias{delete.duplicates}
\title{Delete Duplicates from Modules}
\usage{
delete.duplicates(nodes, modules, graph)
}
\arguments{
\item{graph}{}
}
\description{
Delete Duplicates from Modules
}
\keyword{internal}
