% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.supergraphs.R
\name{delete.supergraphs}
\alias{delete.supergraphs}
\title{Delete Supergraphs from Modules}
\usage{
delete.supergraphs(nodes, modules, graph)
}
\arguments{
\item{graph}{}
}
\description{
Delete Supergraphs from Modules
}
\keyword{internal}
