% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDuplicate.R
\name{deleteDuplicates}
\alias{deleteDuplicates}
\title{Delete Duplicates from Modules}
\usage{
deleteDuplicates(nodes, modules, graph)
}
\arguments{
\item{graph}{}
}
\description{
Delete Duplicates from Modules
}
\keyword{internal}
