% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteSupergraphs.R
\name{deleteSupergraphs}
\alias{deleteSupergraphs}
\title{Delete Supergraphs from Modules}
\usage{
deleteSupergraphs(nodes, modules, graph)
}
\arguments{
\item{graph}{}
}
\description{
Delete Supergraphs from Modules
}
\keyword{internal}
