% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.network.with.modules.R
\name{draw.network.with.modules}
\alias{draw.network.with.modules}
\title{Draw Network with Modules}
\usage{
draw.network.with.modules(graph, title, nodes, colors, save.image=TRUE, cy)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object, which can be generated by
\code{\link{generateNetwork}}.}

\item{title}{a name for the Cytoscape session.}

\item{nodes}{\code{\link[data.table]{data.table}} with data containing the nodes.
It is contained in the output list of \code{\link{identifyModules}}.}

\item{colors}{a colour palette as returned from 
\code{\link[grDevices]{rainbow}} for colouring the different modules.}

\item{save.image}{TRUE, if modules should be saved as png files.}

\item{cy}{a \code{\link[RCytoscape]{CytoscapeConnectionClass-class}} object, 
which can be created with
\code{\link[RCytoscape]{CytoscapeConnection}}.}
}
\description{
Draw Network with Modules
}
\references{
\insertRef{Shannon2003}{MoDentify}

\insertRef{Smoot2011}{MoDentify}

\insertRef{Bot2011}{MoDentify}

\insertRef{Shannon2013}{MoDentify}
}
