% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawModules.R
\name{drawModules}
\alias{drawModules}
\title{Draw Modules}
\usage{
drawModules(graph, summary, title="",
close.cycnets.afterwards=FALSE, save.image=FALSE, modules.to.draw=NULL,
only.overview.network = TRUE)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object, which can be generated by
\code{\link{generateNetwork}}.}

\item{summary}{a list containing the modules and nodes}

\item{title}{a name for the Cytoscape session.}

\item{close.cycnets.afterwards}{TRUE, if the windows in the cytoscape environment should be closed
after drawing. This might be useful repeated function call.}

\item{save.image}{TRUE, if modules should be saved as png files.}

\item{modules.to.draw}{a vector containing the IDs of the modules, that should be drawn.
If \code{NULL}, all modules will be drawn.}

\item{only.overview.network}{TRUE if only the overall network should be drawn,
but not the individual networks for the modules}
}
\description{
Draw Modules
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)
data(qmdiab.phenos)

data <- qmdiab.data[, 1:75]
annotations <- qmdiab.annos[1:75]

net_graph <- generateNetwork(data = data, annotations = annotations)
mods <- identifyModules(
  graph = net_graph, data = data, annotations = annotations,
  phenotype = qmdiab.phenos$T2D
)
\donttest{
drawModules(
  graph = net_graph, summary = mods, title = "modules",
  save.image = FALSE
)
}

}
\references{
\insertRef{Shannon2003}{MoDentify}

\insertRef{Smoot2011}{MoDentify}

\insertRef{RCy3}{MoDentify}
}
