% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.modules.R
\name{export.modules}
\alias{export.modules}
\title{Export Modules}
\usage{
export.modules(modules, file)
}
\arguments{
\item{modules}{a list as returned by \code{\link{identify.modules}}}

\item{file}{a name for the txt file that should be stored}
}
\description{
Export Modules
}
