% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNetwork.R
\name{generateNetwork}
\alias{generateNetwork}
\title{Network inference with Gaussian graphical models (GGMs) or Pearson correlation}
\usage{
generateNetwork(data, covars = NULL, annotations,
  correlation.type = "partial", alpha = 0.05,
  correction.method = "bonferroni")
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}} or matrix containing the data.
The columns correspond to the variables (e.g. metabolites), the rows to the observations.}

\item{covars}{a \code{\link[data.table]{data.table}} containing covariates to correct for.
The columns correspond to the different covariates, the rows to the observations.}

\item{annotations}{a \code{\link[data.table]{data.table}} containing annotations for the variables (e.g. pathway annotations).
The columns correspond to the different annotations, the rows to the variables}

\item{correlation.type}{type of correlation to be estimated. Can either be "pearson", or "partial".}

\item{alpha}{significance level (type 1 error) for multiple testing correction.}

\item{correction.method}{the method that should be used for multiple testing correction ("bonferroni", "BH", "BY", "fdr", "holm", "hochberg", "hommel", "none").
Default is bonferroni. See \code{\link[stats]{p.adjust}}.}
}
\value{
a network containing the variables as nodes as an \code{\link[igraph]{igraph}} object.
}
\description{
Network inference with Gaussian graphical models (GGMs) or Pearson correlation
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)

net.graph<-generateNetwork(data = qmdiab.data, annotations = qmdiab.annos, 
alpha=0.05, correction.method = "bonferroni")
}
\references{
\insertRef{Krumsiek2011}{MoDentify}
}
