% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePathwaysNetwork.R
\name{generatePathwaysNetwork}
\alias{generatePathwaysNetwork}
\title{Subnetwork Creation function}
\usage{
generatePathwaysNetwork(data, covars = NULL, annotations,
  level = "Sub.pathway", correlation.type = "partial", alpha = 0.05,
  correction.method = "bonferroni", rm.unknown = TRUE,
  representative.method = "eigenmetabolite")
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}} or matrix containing data.
The columns correspond to the variables, the rows to the observations.}

\item{covars}{a \code{\link[data.table]{data.table}} containing covariates to correct for.
The columns correspond to the different covariates, the rows to the observations.}

\item{annotations}{a \code{\link[data.table]{data.table}} containing annotations for the variables.
The columns correspond to the different annotations, the rows to the variables.}

\item{level}{name of the column that should be used for grouping the variables. The default is "SUB_PATHWAY".}

\item{alpha}{significance threshold (type 1 error) for multiple testing correction.}

\item{correction.method}{the method that should be used for multiple testing correction ("bonferroni", "BH", "BY", "fdr", "holm", "hochberg", "hommel", "none").
Default is bonferroni. See \code{\link[stats]{p.adjust}}.}

\item{rm.unknown}{remove variables, for which the group is unknown. The corresponding grouping variable name should be "Unknown".
Default is \code{TRUE}.}

\item{representative.method}{the method, that is used for calculation of the eigenmetabolites.
Currently implemented: "eigenmetabolite" and "average"}
}
\value{
a list with a network containing the variables as nodes as
an \code{\link[igraph]{igraph}} object and a list with the module 
representatives including their explained variances.
}
\description{
Subnetwork Creation function
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)

pathway.graph<-generatePathwaysNetwork(data = qmdiab.data,
 annotations = qmdiab.annos, level = "Sub.pathway")
}
\references{
\insertRef{Krumsiek2011}{MoDentify}

\insertRef{Do2017}{MoDentify}
}
