% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.merged.modules.R
\name{get.merged.modules}
\alias{get.merged.modules}
\title{Get Module scores for Merged Modules}
\usage{
get.merged.modules(graph, data, phenotype, covars, nodes)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object, which can be generated with \code{\link{create.network}}.
The ID of the nodes must correspond to the name of the variables.}

\item{data}{either a matrix, where the columns correspond to the variables and the rows to the observations.
Or a \code{\link[data.table]{data.table}} with three columns: name, sampleID and value.}

\item{phenotype}{a vector with the values for a phenotype of interest.
It must have the same number of observations as in data.}

\item{covars}{a \code{\link[data.table]{data.table}} containing the covariates to correct for.}

\item{nodes}{an \code{\link[igraph]{igraph}} object, containing the information for the nodes and their
new modules.}
}
\description{
Get Module scores for Merged Modules
}
