% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.neighbors.for.module.R
\name{get.neighbors.for.module}
\alias{get.neighbors.for.module}
\title{Get Neighbors for Module}
\usage{
get.neighbors.for.module(graph, module)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object}

\item{module}{a vector containing nodes corresponding to the graph}
}
\value{
a vector containing all neighbors of the module
}
\description{
Get Neighbors for Module
}
\examples{
net.graph<-generate.network(data=qmdiab.data, annotations=qmdiab.annos)
module<-c(3,4)
neighbors<-get.neighbors.for.module(graph = net.graph, module = module)
}
\keyword{internal}
