% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighborsForModule.R
\name{getNeighborsForModule}
\alias{getNeighborsForModule}
\title{Get Neighbors for Module}
\usage{
getNeighborsForModule(graph, module)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object}

\item{module}{a vector containing nodes corresponding to the graph}
}
\value{
a vector containing all neighbors of the module
}
\description{
Get Neighbors for Module
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)

net.graph<-generateNetwork(data=qmdiab.data, annotations=qmdiab.annos)
module<-c(3,4)
neighbors<-getNeighborsForModule(graph = net.graph, module = module)
}
\keyword{internal}
