% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.edges.between.attributes.R
\name{gg.edges.between.attributes}
\alias{gg.edges.between.attributes}
\title{Plotting of the edges between attribute types}
\usage{
gg.edges.between.attributes(graph, name, rm.unknown = FALSE)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object}

\item{name}{the name of the attribute}
}
\description{
Plotting of the edges between attribute types
}
\examples{
net.graph<-generate.network(data = qmdiab.data, annotations = qmdiab.annos)
gg.edges.between.attributes(net.graph, "Super.pathway")
}
