% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.explained.variance.R
\name{gg.explained.variance}
\alias{gg.explained.variance}
\title{Plotting of the explained variances}
\usage{
gg.explained.variance(subnet, selection = NULL)
}
\arguments{
\item{subnet}{a list generated by the \code{\link{generate.pathways.network}} function}
}
\description{
Plotting of the explained variances
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)

pathway.graph<-generate.pathways.network(data = qmdiab.data,
 annotations = qmdiab.annos, level = "Sub.pathway")
 gg.explained.variance(pathway.graph)
}
