% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggEdgesBetweenAttributes.R
\name{ggEdgesBetweenAttributes}
\alias{ggEdgesBetweenAttributes}
\title{Plotting of the edges between attribute types}
\usage{
ggEdgesBetweenAttributes(graph, name, rm.unknown = FALSE)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object}

\item{name}{the name of the attribute}

\item{rm.unknown}{remove unknowns}
}
\description{
Plotting of the edges between attribute types
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)

net.graph<-generateNetwork(data = qmdiab.data, annotations = qmdiab.annos)
ggEdgesBetweenAttributes(net.graph, "Super.pathway")
}
