% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExplainedVariance.R
\name{ggExplainedVariance}
\alias{ggExplainedVariance}
\title{Plotting of the explained variances}
\usage{
ggExplainedVariance(subnet, selection = NULL)
}
\arguments{
\item{subnet}{a list generated by the \code{\link{generatePathwaysNetwork}} function}

\item{selection}{which groups should be used for the plot.}
}
\description{
Plotting of the explained variances
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)

pathway.graph<-generatePathwaysNetwork(data = qmdiab.data,
 annotations = qmdiab.annos, level = "Sub.pathway")
 ggExplainedVariance(pathway.graph)
}
