% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.modules.R
\name{identify.modules}
\alias{identify.modules}
\title{Module Identification}
\usage{
identify.modules(graph, data, phenotype, covars = NULL, annotations,
merge.overlapping=FALSE, better.than.components= TRUE, alpha=0.05,
for.pathways=FALSE, representative.method="mean", correction.method="bonferroni")
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object, which can be generated with \code{\link{create.network}}.
The ID of the nodes must correspond to the name of the variables.}

\item{data}{either a matrix, where the columns correspond to the variables and the rows to the observations.
Or a \code{\link[data.table]{data.table}} with three columns: name, sampleID and value.}

\item{phenotype}{a vector with the values for a phenotype of interest.
It must have the same number of observations as data.}

\item{covars}{a \code{\link[data.table]{data.table}} containing the covariates
to correct for.
The rows for the samples must be in the same order as in the phenotype vector.}

\item{annotations}{a \code{\link[data.table]{data.table}} containing annotations for the variables.
The columns correspond to the different annotations, the rows to the variables.}

\item{merge.overlapping}{if \code{TRUE}, overlapping modules will be merged.}

\item{better.than.components}{if \code{TRUE}, modules will only be enlarged and accepted,
if they are better than all of their components.}

\item{alpha}{significance level for accepting the modules.}

\item{for.pathways}{Must be set to the name of the column to be used, if modules should be calculated for pathways.}

\item{representative.method}{the method, that is used for the calculation of the module representation.
Currently implemented: "eigenM-PCA", "eigenM-SVD" and "mean"}

\item{correction.method}{the method that used for multiple testing correction ("bonferroni", "BH", "BY", "fdr", "holm", "hochberg", "hommel", "none").
Default is set to bonferroni. See \code{\link[stats]{p.adjust}}.}
}
\value{
a list consisting of four elements.
}
\description{
Module Identification
}
\examples{
data<-qmdiab.data[, 1:75]
annotations<-qmdiab.annotations[1:75]

net.graph<-create.network(data=data, annotations=annotations)
mods<-identify.modules(graph=net.graph, data=data, phenotype = qmdiab.phenotype$T2D,
alpha = 0.05)

pathway.graph<-create.pathways.network(data=data, annotations=annotations)$network
pathway.mods<-identify.modules(graph=pathway.graph, data=data,
phenotype = qmdiab.phenotype$T2D, for.pathways = TRUE, annotations = annotations,
alpha = 0.05)
}
\references{
\insertRef{Do2017}{MoDentify}

\insertRef{Chuang2007}{MoDentify}
}
