% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearScoring.R
\name{linearScoring}
\alias{linearScoring}
\title{Linear Scoring Function}
\usage{
linearScoring(moduleRepresentatives, phenotype, covars)
}
\arguments{
\item{moduleRepresentatives}{a \code{\link[data.table]{data.table}} containing the module
representative for each sample as returned by \code{\link[MoDentify]{calculateModuleRepresentatives}}}

\item{phenotype}{a vector with the values for a phenotype of interest.
It must have the same number of observations as in data.}

\item{covars}{a \code{\link[data.table]{data.table}} containing the covariates.
The rows for the observations must be in the same order as in the phenotype vector.}
}
\value{
a list containing the module score and the regression coefficient
}
\description{
scores the module in regard to a given phenotype using a linear regression
}
\examples{
#TODO
}
\seealso{
\code{\link[stats]{lm}}
}
