% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoDentify.R
\docType{package}
\name{MoDentify-package}
\alias{MoDentify}
\alias{MoDentify-package}
\title{Phenotype-driven module identification}
\description{
Phenotype-driven module identification
}
\references{
\insertRef{Do2017}{MoDentify}
}
\author{
\strong{Maintainer}: David Rasp \email{david.j.rasp@gmail.com}, Kieu Trinh Do \email{kieu.do@helmholtz-muenchen.de}

Authors:
\itemize{
  \item Kieu Trinh Do \email{kieu.do@helmholtz-muenchen.de}
  \item Gabi Kastenmueller \email{g.kastenmueller@helmholtz-muenchen.de}
  \item Karsten Suhre \email{karsten@suhre.de}
  \item Jan Krumsiek \email{jan.krumsiek@helmholtz-muenchen.de}
}

}
