% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.annotation.R
\name{order.annotation}
\alias{order.annotation}
\title{Order Annotations}
\usage{
order.annotation(annot_subnet_DT, names)
}
\arguments{
\item{names}{}
}
\description{
Order Annotations
}
\keyword{internal}
