% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoDentify.R
\docType{data}
\name{qmdiab.annos}
\alias{qmdiab.annos}
\title{qmdiab annotations}
\format{a \code{\link[data.table]{data.table}} with 1524 rows and 14 variables}
\usage{
data(qmdiab.annos)
}
\description{
qmdiab annotations
}
\keyword{datasets}
