% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoDentify.R
\docType{data}
\name{qmdiab.data}
\alias{qmdiab.data}
\title{qmdiab data}
\format{a \code{\link[data.table]{data.table}} with 310 rows and 1524 colums}
\usage{
data(qmdiab.data)
}
\description{
qmdiab data
}
\keyword{datasets}
