% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoDentify.R
\docType{data}
\name{qmdiab.phenos}
\alias{qmdiab.phenos}
\title{qmdiab phenotypes}
\format{a \code{\link[data.table]{data.table}} with 310 rows and 4 columns}
\usage{
data(qmdiab.phenos)
}
\description{
qmdiab phenotypes
}
\keyword{datasets}
