% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representativeAverage.R
\name{representativeAverage}
\alias{representativeAverage}
\title{Average Level as Module Representative}
\usage{
representativeAverage(data)
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}}, containg the values of the current module,
with three columns: name, sampleID and z-score.}
}
\description{
Average Level as Module Representative
}
\references{
\insertRef{Do2017}{MoDentify}

\insertRef{Chuang2007}{MoDentify}
}
