% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representativeEigenMetabolite.R
\name{representativeEigenMetabolite}
\alias{representativeEigenMetabolite}
\title{Eigen Metabolite as Module Representative}
\usage{
representativeEigenMetabolite(data)
}
\arguments{
\item{data}{a \code{\link[data.table]{data.table}}, containg the values of the current module,
with three columns: name, sampleID and z-score.}
}
\description{
Eigen Metabolite as Module Representative
}
\references{
\insertRef{Do2017}{MoDentify}

\insertRef{Langfelder2007}{MoDentify}
}
