% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{volcano.plot}
\alias{volcano.plot}
\title{Volcano Plot}
\usage{
volcano.plot(modules, data, phenotype, for.modules = FALSE,
  representative.method = "mean")
}
\arguments{
\item{modules}{a list as returned by \code{\link{identify.modules}}}

\item{data}{a \code{\link[data.table]{data.table}} or matrix containing data.
The columns correspond to the variables, the rows to the observations.}

\item{phenotype}{a vector with the values for a phenotype of interest.
It must have the same number of samples as data.}

\item{for.modules}{If volcano plots should be generated for the modules.}

\item{representative.method}{the method that is used for the calculation of the representative.
Currently implemented: "eigenM-PCA", "eigenM-SVD" and "mean". Only needed if
volcano plots should be drawn for the modules and should be the same as used in the
module identification.}
}
\description{
Volcano Plot
}
\examples{
data<-simulated.data[, 1:75]
annotations<-simulated.annotations[1:75]
net.graph<-create.network(data=data, annotations=annotations)
modules<-identify.modules(graph=net.graph, data=data, phenotype = simulated.phenotype)
volcano.plot(modules=modules, data=data, phenotype=simulated.phenotype)
}
