% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModules.R
\name{writeModules}
\alias{writeModules}
\title{Export Modules}
\usage{
writeModules(modules, file)
}
\arguments{
\item{modules}{a list as returned by \code{\link{identify.modules}}}

\item{file}{a name for the txt file that should be stored}
}
\description{
Export Modules
}
