% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.modules.R
\name{draw.modules}
\alias{draw.modules}
\title{Draw Modules}
\usage{
draw.modules(graph, title="", nodes, modules,
close.cycnets.afterwards=FALSE, save.image=TRUE, modules.to.draw=NULL)
}
\arguments{
\item{graph}{an \code{\link[igraph]{igraph}} object, which can be generated by
\code{\link{create.network}}.}

\item{title}{a name for the Cytoscape session.}

\item{nodes}{a \code{\link[data.table]{data.table}} with data containing the nodes.
It is contained in the output list of \code{\link{identify.modules}}.}

\item{modules}{a \code{\link[data.table]{data.table}} with data containing the nodes in the modules.
It is contained in the output list of \code{\link{identify.modules}}.}

\item{close.cycnets.afterwards}{TRUE, if the windows in the cytoscape environment should be closed
after drawing. This might be useful repeated function call.}

\item{save.image}{TRUE, if modules should be saved as png files.}

\item{modules.to.draw}{a vector containing the IDs of the modules, that should be drawn.
If \code{NULL}, all modules will be drawn.}
}
\description{
Draw Modules
}
\examples{
data<-qmdiab.data[, 1:75]
annotations<-qmdiab.annotations[1:75]
net_graph<-create.network(data=data, annotations=annotations)
mods<-identify.modules(graph=net_graph, data=data, phenotype = qmdiab.phenotype$T2D)
\donttest{draw.modules(graph=net_graph, title="modules", nodes = mods$nodes,
modules = mods$modules, save.image=FALSE)}

}
\references{
\insertRef{Shannon2003}{MoDentify}

\insertRef{Smoot2011}{MoDentify}

\insertRef{Bot2011}{MoDentify}

\insertRef{Shannon2013}{MoDentify}
}
