% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.qmdiab.data.R
\name{get.qmdiab.data}
\alias{get.qmdiab.data}
\title{Calculate eigenmetabolites as module representatives}
\usage{
get.qmdiab.data(filename)
}
\arguments{
\item{url}{the link to where QMDiab data is stored.}
}
\value{
list of three, data: \code{\link[data.table]{data.table}} of 
preprocssed metabolite concentrations, annos: 
\code{\link[data.table]{data.table}} of annotations for each metabolite, 
phenos: \code{\link[data.table]{data.table}} of age, sex, BMI, and T2D information
}
\description{
Calculate eigenmetabolites as module representatives
}
\references{
\insertRef{donetwork2015}{MoDentify}
\insertRef{mookkanamori2013}{MoDentify}
\insertRef{suhreconnecting2017}{MoDentify}
\insertRef{yousri2015}{MoDentify}
}
