% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.explained.variance.R
\name{gg.explained.variance}
\alias{gg.explained.variance}
\title{Plotting of the explained variances}
\usage{
gg.explained.variance(subnet)
}
\arguments{
\item{subnet}{a list generated by the \code{\link{create.pathways.network}} function}
}
\description{
Plotting of the explained variances
}
\examples{
pathway.graph<-create.pathways.network(data = qmdiab.data,
 annotations = qmdiab.annotations, column.name = "Sub.pathway")
 gg.explained.variance(pathway.graph)
}
