% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy.module.selection.R
\name{greedy.module.selection}
\alias{greedy.module.selection}
\title{Greedy Module Selection for a Seed}
\usage{
greedy.module.selection(graph, nodeNr, data, phenotype, covars = NULL,
  alpha = 0.05, already_calculated, better.than.components = TRUE,
  representative.method = "average")
}
\arguments{
\item{graph}{an igraph object, which can be generated with \code{\link{generate.network}}.
The ID of the nodes must correspond to the name of the variables.}

\item{nodeNr}{the number of the node, which should be used as a seed.}

\item{data}{a \code{\link[data.table]{data.table}} with three columns: name,
sampleID and z-score.}

\item{phenotype}{a vector with the values for a phenotype of interest.
It must have the same number of samples as in data.}

\item{covars}{a \code{\link[data.table]{data.table}} containing the covariates to correct for.
The rows for the observations must be in the same order as in the phenotype vector.}

\item{alpha}{significance level (type 1 error) for accepting the modules.}

\item{already_calculated}{a \code{\link[data.table]{data.table}} containing the results for
already calculated modules}

\item{better.than.components}{if \code{TRUE}, modules will only be enlarged and accepted,
if they are better than all of their components.}

\item{representative.method}{the method used for the calculation of the module representatives.
Currently implemented: "eigenmetabolite" and "average"}
}
\value{
a list containing the members of the module, the module-score, its regression coefficient
for the given phenotype, the score and regression coefficient for the seed, the cache
and the consecutive module-scores after adding each new node.
}
\description{
Greedy Module Selection for a Seed
}
\examples{
data(qmdiab.data)
data(qmdiab.annos)
data(qmdiab.phenos)

net.graph<-generate.network(data=qmdiab.data, annotations=qmdiab.annos)
data<-data.table(sampleID= paste0("sample", 1:dim(qmdiab.data)[1]),
qmdiab.data)
data<-melt(data=data, id.vars = "sampleID", variable.name = "name")
data[, z.score:=scale(value), by=.(name)]
already_calculated<-data.table(key.value=character(),
score=numeric(), beta=numeric(), times.accessed=numeric())

module<-greedy.module.selection(graph=net.graph, nodeNr = 51, data = data,
phenotype = qmdiab.phenos$T2D, already_calculated = already_calculated)
}
\references{
\insertRef{Do2017}{MoDentify}
}
