% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoDentify.R
\docType{package}
\name{MoDentify-package}
\alias{MoDentify}
\alias{MoDentify-package}
\title{Phenotype-driven module identification}
\description{
A package for the identification of functional modules, that are associated with a given phenotype, in networks
}
\references{
\insertRef{Do2017}{MoDentify}

\insertRef{Do2018}{MoDentify}
}
\author{
\strong{Maintainer}: Jan Krumsiek \email{jan.krumsiek@helmholtz-muenchen.de}

Authors:
\itemize{
  \item Kieu Trinh Do \email{kieu.do@helmholtz-muenchen.de}
  \item TEST TEST \email{test@helmholtz-muenchen.de}
  \item David Rasp \email{david.j.rasp@gmail.com}
  \item Gabi Kastenmueller \email{g.kastenmueller@helmholtz-muenchen.de}
}

}
