% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoDentify.R
\docType{package}
\name{MoDentify-package}
\alias{MoDentify}
\alias{MoDentify-package}
\title{Phenotype-driven module identification}
\description{
MoDentify is a package for the identification of functional modules, 
            that are associated with a given phenotype, in networks. It was 
            developed for the use with metabolic networks, but could be applied
            to other biological networks as well.
}
\references{
\insertRef{Do2017}{MoDentify}

\insertRef{Do2018}{MoDentify}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/David-J-R/MoDentify}
  \item Report bugs at \url{https://github.com/David-J-R/MoDentify/issues}
}

}
\author{
\strong{Maintainer}: David Rasp \email{david.j.rasp@gmail.com}

Authors:
\itemize{
  \item Kieu Trinh Do \email{kieu.do@helmholtz-muenchen.de}
  \item Gabi Kastenmueller \email{g.kastenmueller@helmholtz-muenchen.de}
  \item Karsten Suhre \email{karsten@suhre.fr}
  \item Jan Krumsiek \email{jan.krumsiek@helmholtz-muenchen.de}
}

}
